% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probly.R
\name{cards}
\alias{cards}
\title{Creates histogram, boxplot and numeric summary}
\usage{
cards(jokers = FALSE, makespace = FALSE)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Creates histogram, boxplot and numeric summary
}
