% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{extract_cor_cov_samps}
\alias{extract_cor_cov_samps}
\title{Extract correlation/covariance matrix samples}
\usage{
extract_cor_cov_samps(splt_fit, par_subscript = "ep")
}
\arguments{
\item{splt_fit}{a fit of class stanfit}

\item{par_subscript}{the subscript for the L_Omega_subscript and tau_subscript parameters}
}
\value{
a list with both Omega (correlation matrix) and Sigma (covariance matrix) sample arrays.
}
\description{
Because of the decomposition of the beta covariance matrix into scale and
correlation, it can be hard to interpret the posteriors. This function will
return an array which is a T-by-T-by-Samples array (where T is the number of
coefficients). It is then easy to use apply to get posterior means for each
cell in the matrix (e.g., \code{apply(Sigma_Array, c(1, 2), mean)}).
}
