% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_col_as_trial_matrix}
\alias{get_col_as_trial_matrix}
\title{Get column as a trial matrix}
\usage{
get_col_as_trial_matrix(splt_df, col, id_col = "id", sample_col = "sample",
  trial_col = "trial_index")
}
\arguments{
\item{splt_df}{a data frame containing data from the SPLT}

\item{col}{the column to use for cell values in the returned matrix}

\item{id_col}{the column if individual IDs}

\item{sample_col}{the column denoting the sample for each row}

\item{trial_col}{the column indexing trials, to ensure correct order}
}
\value{
an N-individual by T-trial matrix with values taken from the column
  in \code{splt_df} named by \code{col}
}
\description{
Create an N-individual by T-trial matrix from a column in a SPLT data frame
}
