% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_date_from_epoch_ms}
\alias{get_date_from_epoch_ms}
\title{Get date from ms since epoch}
\usage{
get_date_from_epoch_ms(miliseconds_since_epoch)
}
\arguments{
\item{miliseconds_since_epoch}{how many miliseconds have passed since 1970-01-01}
}
\value{
a date
}
\description{
Get date from ms since epoch
}
