% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_max_trials_per_individual}
\alias{get_max_trials_per_individual}
\title{Get max trials per individual}
\usage{
get_max_trials_per_individual(i_by_t_mat)
}
\arguments{
\item{i_by_t_mat}{a matrix of N-individuals by T-trials, with NA in any cells
after that individuals max number of trials}
}
\value{
a vector of the number of trials for each individual
}
\description{
Easy creation of a vector with the number of trials for each individual
}
