% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{get_par_summaries}
\alias{get_par_summaries}
\title{Get par summaries}
\usage{
get_par_summaries(afit, par_regex, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{afit}{an rstan fit}

\item{par_regex}{regular expression for \code{grep} to use to get paramater names}
}
\value{
a table of summary statistics from all chains in \code{afit}
}
\description{
Get par summaries
}
