% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_sample_index}
\alias{get_sample_index}
\title{Get sample index}
\usage{
get_sample_index(splt_df, id_col = "id", sample_col = "sample",
  levels = sort(unlist(unique(splt_df[, sample_col]))))
}
\arguments{
\item{splt_df}{a data frame containing data from the SPLT}

\item{id_col}{the name of the id column}

\item{sample_col}{the name of the sample column}

\item{levels}{the levels to be used in coercing the sample column to a factor
(to control ordering)}
}
\value{
a data frame with columns id_col, sample_col, m_fac (with sample_col
  coerced to a factor) and m (with m_fac coerced to a number)
}
\description{
Creates a data frame that links each individual to a particular sample. The
column \code{m} can be easily passed to the data simulation functions, and to
the Stan reinforcement learning model.
}
