% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{make_task_structure_from_data}
\alias{make_task_structure_from_data}
\title{Make task structure from data}
\usage{
make_task_structure_from_data(splt_df)
}
\arguments{
\item{splt_df}{a data frame from the SPLT, with non-response (\code{is.na(pressed_r)}) rows removed}
}
\value{
a list containing the elements defining behavior from a SPLT sample
  from which to simulate data.
}
\description{
Make task structure from data
}
