% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{plot_splt_sim_behavior}
\alias{plot_splt_sim_behavior}
\title{Plot simulated SPLT behavior}
\usage{
plot_splt_sim_behavior(splt_df, press_right_mat)
}
\arguments{
\item{splt_df}{a data frame from the SPLT, with non-response (\code{is.na(pressed_r)}) rows removed}

\item{press_right_mat}{an individual x trial matrix of trial responses}
}
\value{
a plot split by sample, and overall summary plot
}
\description{
Plot simulated SPLT behavior
}
