% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{sample_deltas}
\alias{sample_deltas}
\title{Sample Deltas}
\usage{
sample_deltas(mu_vec, sigma, nsamples)
}
\arguments{
\item{mu_vec}{vector of means of length K (one mean per condition)}

\item{sigma}{positive real number for standard deviation of means across
samples}

\item{nsamples}{number of samples over which deltas should vary}
}
\value{
deltas is a M = nsamples by K condition matrix
}
\description{
This will provide mean coefficients for each sample that you can use to
construct individual-level coefficients.
}
