% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tools.R
\name{simulate_splt_data}
\alias{simulate_splt_data}
\title{Simulate SPLT data}
\usage{
simulate_splt_data(splt_structure, mu_xi, mu_b, mu_eps, mu_rho)
}
\arguments{
\item{splt_structure}{a list, as returned from
\code{\link{make_task_structure_from_data}}}

\item{mu_xi}{a vector of length splt_structure$K of population means for \eqn{\xi}}

\item{mu_b}{a vector of length splt_structure$K of population means for \eqn{b}}

\item{mu_eps}{a vector of length splt_structure$K of population means for
\eqn{\epsilon}}

\item{mu_rho}{a vector of length splt_structure$K of population means for
\eqn{\rho}}
}
\value{
a list with simulated behavior where element \code{press_right} is 0
  if the simulation chose 'left' and 1 if it chose 'right'; and where element
  \code{outcome_realized} contains the value for each trial that resulted
  from the choice.
}
\description{
Simulate SPLT data
}
