% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splt}
\alias{splt}
\title{Social Probabilistic Learning Task Data}
\format{A data frame with M rows and N variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\usage{
splt
}
\description{
Description
}
\keyword{datasets}
