% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splt_dev_and_demog}
\alias{splt_dev_and_demog}
\title{Social Probabilistic Learning Development and Demographic Data}
\format{A data frame with M rows and N variables:
\describe{
  \item{blah}{blah}
}}
\usage{
splt_dev_and_demog
}
\description{
Description
}
\keyword{datasets}
