% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splt_fsmi}
\alias{splt_fsmi}
\title{Fundamental Social Motives Inventory Data}
\format{A data frame with M rows and N variables: \describe{
  \item{blah}{blah} }}
\usage{
splt_fsmi
}
\description{
College sample only. Questionnaire source is:
}
\details{
Neel, R., Kenrick, D. T., White, A. E., & Neuberg, S. L. (2015). Individual
Differences in Fundamental Social Motives. Journal of Personality and Social
Psychology, No Pagination Specified. https://doi.org/10.1037/pspp0000068
}
\keyword{datasets}
