% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tools.R
\name{CachedFit}
\alias{CachedFit}
\title{Loads previous model fit, or fits and saves}
\usage{
CachedFit(expr, rds_filename)
}
\arguments{
\item{expr}{An expression (e.g., \code{\{lm(y ~ x, data = aDataFrame)\}}) to be evaluated.}

\item{rds_filename}{A filename (full path, or filename in current working directory) to save or load using \code{saveRDS} or \code{loadRDS}.}
}
\value{
the return value of \code{expr} or the object saved in the RDS object located at \code{rds_filename}
}
\description{
\code{CachedFit} takes an expression to evaluate (presumably a model fitting call, but it could be anything), and the filename to which to save the fitted model. If the RDS object exists, it loads it instead of fitting the object. It does not check to make sure that the expression corresponds to that which resulted in the original model fit.
}
