% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splt_confidence}
\alias{splt_confidence}
\title{Social Probabilistic Learning Confidence Data}
\format{A data frame with M rows and N variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
splt_confidence
}
\description{
Description
}
\keyword{datasets}
